[codeshuffle] - moved AutoPool, DarwinUtils and GNUOsxReplacements to darwin platform
