+++ tools/darwin/Support/CopyRootFiles-ios.command
@@ -31,7 +31,7 @@ mkdir -p "$TARGET_BUILD_DIR/$TARGET_NAME/AppData/AppHome/media"
 mkdir -p "$TARGET_BUILD_DIR/$TARGET_NAME/AppData/AppHome/tools/darwin/runtime"
 
 ${SYNC} "$SRCROOT/LICENSE.GPL"  "$TARGET_BUILD_DIR/$TARGET_NAME/AppData/"
-${SYNC} "$SRCROOT/xbmc/osx/Credits.html"  "$TARGET_BUILD_DIR/$TARGET_NAME/AppData/"
+${SYNC} "$SRCROOT/xbmc/platform/darwin/Credits.html"  "$TARGET_BUILD_DIR/$TARGET_NAME/AppData/"
 ${ADDONSYNC} "$SRCROOT/addons"  "$TARGET_BUILD_DIR/$TARGET_NAME/AppData/AppHome"
 ${SYNC} "$SRCROOT/media"    "$TARGET_BUILD_DIR/$TARGET_NAME/AppData/AppHome"
