+++ xbmc/cores/VideoPlayer/DVDCodecs/Video/DVDVideoCodecVDA.cpp
@@ -33,7 +33,7 @@ extern "C" {
 #include "cores/VideoPlayer/DVDCodecs/DVDCodecUtils.h"
 #include "cores/FFmpeg.h"
 #include "osx/CocoaInterface.h"
-#include "osx/DarwinUtils.h"
+#include "platform/darwin/DarwinUtils.h"
 #include "windowing/WindowingFactory.h"
 #include "utils/BitstreamConverter.h"
 #include "utils/log.h"
