+++ xbmc/cores/VideoPlayer/DVDCodecs/Video/VDA.cpp
@@ -20,7 +20,7 @@
 #include "system.h"
 #ifdef TARGET_DARWIN_OSX
 #include "osx/CocoaInterface.h"
-#include "osx/DarwinUtils.h"
+#include "platform/darwin/DarwinUtils.h"
 #include "DVDVideoCodec.h"
 #include "DVDCodecs/DVDCodecUtils.h"
 #include "utils/log.h"
