+++ xbmc/cores/VideoPlayer/VideoRenderers/HwDecRender/RendererIMX.cpp
@@ -28,7 +28,6 @@
 #include "utils/GLUtils.h"
 #include "settings/MediaSettings.h"
 #include "windowing/WindowingFactory.h"
-#include "osx/DarwinUtils.h"
 #include "cores/VideoPlayer/VideoRenderers/RenderCapture.h"
 
 CRendererIMX::CRendererIMX()
