+++ xbmc/cores/VideoPlayer/VideoRenderers/HwDecRender/RendererVTB.cpp
@@ -27,7 +27,7 @@
 #include "utils/GLUtils.h"
 #include "settings/MediaSettings.h"
 #include "windowing/WindowingFactory.h"
-#include "osx/DarwinUtils.h"
+#include "platform/darwin/DarwinUtils.h"
 
 CRendererVTB::CRendererVTB()
 {
