+++ xbmc/cores/VideoPlayer/VideoRenderers/LinuxRendererGL.cpp
@@ -58,7 +58,7 @@ extern "C" {
 #include "osx/CocoaInterface.h"
 #include <CoreVideo/CoreVideo.h>
 #include <OpenGL/CGLIOSurface.h>
-#include "osx/DarwinUtils.h"
+#include "platform/darwin/DarwinUtils.h"
 #endif
 
 //due to a bug on osx nvidia, using gltexsubimage2d with a pbo bound and a null pointer
