+++ xbmc/osx/ios/IOSKeyboard.mm
@@ -22,9 +22,9 @@
 #include "IOSKeyboard.h"
 #include "IOSKeyboardView.h"
 #include "XBMCDebugHelpers.h"
-#include "osx/DarwinUtils.h"
+#include "platform/darwin/DarwinUtils.h"
 
-#import "AutoPool.h"
+#import "platform/darwin/AutoPool.h"
 
 KeyboardView *g_pIosKeyboard = nil;
