+++ xbmc/platform/darwin/AutoPool.h
