+++ xbmc/platform/darwin/DarwinUtils.h
