+++ xbmc/platform/darwin/DarwinUtils.mm
