+++ xbmc/platform/darwin/OSXGNUReplacements.c
