+++ xbmc/platform/darwin/OSXGNUReplacements.h
