+++ xbmc/settings/Settings.cpp
@@ -48,7 +48,7 @@
 #include "osx/XBMCHelper.h"
 #endif // defined(TARGET_DARWIN_OSX)
 #if defined(TARGET_DARWIN)
-#include "osx/DarwinUtils.h"
+#include "platform/darwin/DarwinUtils.h"
 #endif
 #if defined(TARGET_DARWIN_IOS)
 #include "SettingAddon.h"
